Moved ConstantDeadTimeController to devices/controllers.py and updated devices to use (#1465)
* Created devices/controllers.py and updated devices to use

* Fixed tests

* Updated doc string

* Updated doc string

* Updated merlin controller to use ConstantDeadTimeController

* Add parametised test

* made deadtime veriable
