+++ src/dodal/devices/i13_1/merlin_controller.py
@@ -8,15 +8,16 @@ from ophyd_async.core import (
 )
 from ophyd_async.epics.adcore import (
     DEFAULT_GOOD_STATES,
-    ADBaseController,
     ADBaseIO,
     ADImageMode,
     ADState,
 )
 from ophyd_async.epics.core import stop_busy_record
 
+from dodal.devices.controllers import ConstantDeadTimeController
 
-class MerlinController(ADBaseController):
+
+class MerlinController(ConstantDeadTimeController):
     def __init__(
         self,
         driver: ADBaseIO,
@@ -29,8 +30,7 @@ class MerlinController(ADBaseController):
         for drv_child in self.driver.children():
             logging.debug(drv_child)
 
-    def get_deadtime(self, exposure: float | None) -> float:
-        return 0.002
+        super().__init__(driver, 0.002)
 
     async def prepare(self, trigger_info: TriggerInfo):
         self.frame_timeout = (
